create navigation folder and move navigation component to that, add some styling for better visualization
