+++ src/components/Navigation/Navigation.module.css
@@ -1,0 +1,22 @@
+.navContainer {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0 1rem;
+}
+
+.logo {
+  display: flex;
+  align-items: center;
+}
+
+.logo img {
+  width: 100px;
+  height: auto;
+}
+
+.list {
+  display: flex;
+  list-style: none;
+  column-gap: 1rem;
+}
