+++ src/components/Rockets/RocketsPage.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Navigation from '../Navigation';
+import Navigation from '../Navigation/Navigation';
 
 function RocketsPage() {
   return (
